pyproject.toml
[tool]
[tool.pdm]
[tool.pdm.scripts]
start = "python main.py"

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Marco Acierno", email = "marcoaciernoemail@gmail.com"},
]
dependencies = [
    "discord-py>=2.2.3",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.1",
    "redis[hiredis]>=4.5.5",
]
requires-python = ">=3.11"
license = {text = "MIT"}
